requirements.txt
torch>=2.1.0
torchvision>=0.16.0
open_clip_torch>=2.24.0
numpy>=1.24.0
Pillow>=10.0.0
tqdm>=4.66.0
rich>=13.0.0
matplotlib>=3.8.0
seaborn>=0.13.0
umap-learn>=0.5.5
scikit-learn>=1.3.0
jupyter>=1.0.0
ipykernel>=6.0.0
ipywidgets>=8.0.0
